requirements.txt
alto-xml==0.0.5
astroid==2.8.6
autopep8==1.6.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.0
cryptography==36.0.1
distlib==0.3.4
filelock==3.5.0
img2pdf==0.4.3
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.7.1
mypy==0.910
mypy-extensions==0.4.3
pdf2image==1.16.0
pdfminer.six==20211012
Pillow==9.3.0
pipenv==2022.1.8
platformdirs==2.4.1
pycparser==2.21
pylint==2.11.1
pylint-quotes==0.2.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
tesserocr==2.5.2
toml==0.10.2
virtualenv==20.13.1
virtualenv-clone==0.5.7
wrapt==1.13.3
